Use nested repositories where possible.
