--- spring-data-neo4j-rx/src/test/java/org/neo4j/springframework/data/integration/imperative/PersonWithRelationshipWithPropertiesRepository.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2019-2020 "Neo4j,"
- * Neo4j Sweden AB [https://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.neo4j.springframework.data.integration.imperative;
-
-import org.neo4j.springframework.data.integration.shared.PersonWithRelationshipWithProperties;
-import org.neo4j.springframework.data.repository.Neo4jRepository;
-import org.neo4j.springframework.data.repository.query.Query;
-import org.springframework.data.repository.query.Param;
-
-/**
- * @author Gerrit Meier
- */
-public interface PersonWithRelationshipWithPropertiesRepository extends Neo4jRepository<PersonWithRelationshipWithProperties, Long> {
-
-	@Query("MATCH (p:PersonWithRelationshipWithProperties)-[l:LIKES]->(h:Hobby) return p, collect(l), collect(h)")
-	PersonWithRelationshipWithProperties loadFromCustomQuery(@Param("id") Long id);
-}
-
