+++ spring-data-neo4j-rx/src/test/java/org/neo4j/springframework/data/integration/imperative/repositories/ThingRepository.java
@@ -16,18 +16,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.springframework.data.integration.reactive;
+package org.neo4j.springframework.data.integration.imperative.repositories;
 
-import org.neo4j.springframework.data.integration.shared.MultipleLabels;
-import org.neo4j.springframework.data.repository.ReactiveNeo4jRepository;
+import java.util.List;
+
+import org.neo4j.springframework.data.integration.shared.ThingWithAssignedId;
+import org.neo4j.springframework.data.repository.query.Query;
+import org.springframework.data.repository.CrudRepository;
 
 /**
- * @author Gerrit Meier
+ * @author Michael J. Simons
  */
-public interface ReactiveMultipleLabelRepository
-	extends ReactiveNeo4jRepository<MultipleLabels.MultipleLabelsEntity, Long> {
-}
+public interface ThingRepository extends CrudRepository<ThingWithAssignedId, String> {
+	List<ThingWithAssignedId> findFirstByOrderByNameDesc();
 
-interface ReactiveMultipleLabelWithAssignedIdRepository
-	extends ReactiveNeo4jRepository<MultipleLabels.MultipleLabelsEntityWithAssignedId, Long> {
+	List<ThingWithAssignedId> findTop5ByOrderByNameDesc();
+
+	@Query("MATCH (n:Thing{theId:'anId'})-[r:Has]->(b:Thing2) return n, collect(r), collect(b)")
+	ThingWithAssignedId getViaQuery();
 }
