+++ spring-data-neo4j-rx/src/test/java/org/neo4j/springframework/data/integration/reactive/repositories/ReactivePersonRepository.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.neo4j.springframework.data.integration.reactive;
+package org.neo4j.springframework.data.integration.reactive.repositories;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
