+++ spring-data-neo4j-rx/src/test/java/org/neo4j/springframework/data/integration/reactive/repositories/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Repositories shared between tests.
+ */
+package org.neo4j.springframework.data.integration.reactive.repositories;
